requirements.txt
langchain>=0.1.16
langchain-openai>=0.1.3
langchain-community>=0.0.22
langchain-core>=0.1.41
langsmith>=0.1.24

streamlit>=1.32.0
python-dotenv

fastapi>=0.110.0
langserve>=0.0.40
uvicorn[standard]>=0.29.0
sse-starlette
langchain_huggingface
bs4
pypdf
chromadb
faiss-cpu
sentence-transformers